+++ src/ACadSharp/Objects/BookColor.cs
@@ -1,7 +1,6 @@
 ﻿using ACadSharp.Attributes;
 using System;
 using System.Linq;
-using System.Xml.Linq;
 
 namespace ACadSharp.Objects
 {
